# docs/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Brothers App following Clean Architecture.",
        "title": "Brothers App API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/groups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "List my groups",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/group.GroupsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/group.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Create a group",
                "parameters": [
                    {
                        "description": "Group payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/group.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/group.GroupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/group.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/group.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/groups/{groupID}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Owner-only operation. Requires an explicit confirmation and preserves financial and audit history.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Soft-delete a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group UUID",
                        "name": "groupID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Deletion confirmation",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/group.DeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/group.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/group.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/group.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/group.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/groups/{groupID}/invitations": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group invitations"
                ],
                "summary": "Create a group invitation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group UUID",
                        "name": "groupID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Invitation payload; provide user_id or email",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/group.InviteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/group.InvitationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/group.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/group.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/group.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/groups/{groupID}/members": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "List active and pending group members",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group UUID",
                        "name": "groupID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search by name, username, email, or location",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "all",
                            "manager",
                            "employee",
                            "investor"
                        ],
                        "type": "string",
                        "description": "Role filter",
                        "name": "role",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "all",
                            "active",
                            "pending"
                        ],
                        "type": "string",
                        "description": "Status filter",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/group.MembersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/group.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/group.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/group.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/invitations/{invitationID}/action": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group invitations"
                ],
                "summary": "Respond to a group invitation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invitation UUID",
                        "name": "invitationID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Action payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/group.InvitationActionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/group.InvitationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/group.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/group.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/group.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "group.CreateRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "group.DeleteRequest": {
            "type": "object",
            "properties": {
                "confirm": {
                    "type": "boolean"
                }
            }
        },
        "group.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 1400
                },
                "data": {
                    "$ref": "#/definitions/responses.ErrorDetails"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "meta": {
                    "$ref": "#/definitions/responses.Meta"
                },
                "slug": {
                    "type": "string",
                    "example": "invalid_data"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "group.Group": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "is_owner": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "group.GroupListItem": {
            "type": "object",
            "properties": {
                "customers_count": {
                    "type": "integer"
                },
                "group_balance_usd": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "is_owner": {
                    "type": "boolean"
                },
                "locations_count": {
                    "type": "integer"
                },
                "members_count": {
                    "type": "integer"
                },
                "my_profit_uzs": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "order_count": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "subscription_active": {
                    "type": "boolean"
                }
            }
        },
        "group.GroupResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/group.Group"
                },
                "message": {
                    "type": "string",
                    "example": "group created"
                },
                "meta": {
                    "$ref": "#/definitions/responses.Meta"
                },
                "slug": {
                    "type": "string",
                    "example": "ok"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "group.GroupsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/group.GroupListItem"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "groups returned"
                },
                "meta": {
                    "$ref": "#/definitions/responses.Meta"
                },
                "slug": {
                    "type": "string",
                    "example": "ok"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "group.Invitation": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "group_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invited_by": {
                    "type": "string"
                },
                "location_name": {
                    "type": "string"
                },
                "recipient_id": {
                    "type": "string"
                },
                "responded_at": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "group.InvitationActionRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string",
                    "enum": [
                        "accept",
                        "reject"
                    ],
                    "example": "accept"
                }
            }
        },
        "group.InvitationResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/group.Invitation"
                },
                "message": {
                    "type": "string",
                    "example": "invitation created"
                },
                "meta": {
                    "$ref": "#/definitions/responses.Meta"
                },
                "slug": {
                    "type": "string",
                    "example": "ok"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "group.InviteRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "location_name": {
                    "type": "string",
                    "example": "Kokand"
                },
                "role": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "group.Member": {
            "type": "object",
            "properties": {
                "access_level": {
                    "type": "string"
                },
                "avatar_url": {
                    "type": "string"
                },
                "balance_usd": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "invitation_id": {
                    "type": "string"
                },
                "is_owner": {
                    "type": "boolean"
                },
                "location_name": {
                    "type": "string"
                },
                "member_id": {
                    "type": "string"
                },
                "profit_uzs": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "group.MembersResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/group.Member"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "members returned"
                },
                "meta": {
                    "$ref": "#/definitions/responses.Meta"
                },
                "slug": {
                    "type": "string",
                    "example": "ok"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "responses.ErrorDetails": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string",
                    "example": "checking registration email: ERROR: column email does not exist (SQLSTATE 42703)"
                }
            }
        },
...
        "responses.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "example message"
                }
            }
        },
        "responses.Meta": {
            "type": "object",
            "properties": {
                "api_version": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Paste the access token only. Standard clients may send \"Bearer \u003ctoken\u003e\"; Swagger UI sends the token value directly.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}